integrated files

+++ server/files.js
@@ -3,6 +3,20 @@ const fs = require("fs")
 const { MongoClient } = require("mongodb")
 const cryptojs = require("crypto-js")
 const dotenv = require("dotenv")
+const path = require("path")
+const ratelimiter = require("express-rate-limit")
+const limiteroptions = {
+    windowMs: 1000 * 60,
+    limit: 10,
+    statusCode: 200,
+    message: { status: false, message: "Too Many Requests" },
+    legacyHeaders: false,
+    standardHeaders: true
+}
+const uploadlimiter = ratelimiter(limiteroptions)
+const downloadlimiter = ratelimiter(limiteroptions)
+const getfilelimiter = ratelimiter(limiteroptions)
+const codelimiter = ratelimiter(limiteroptions)
 dotenv.config()
 
 module.exports = async function (app) {
@@ -23,15 +37,21 @@ module.exports = async function (app) {
 
     let connect = await MongoClient.connect(process.env.MONGO_URL)
     let db = connect.db("website")
-
+    const updatedownloads = async ()=>{
+        let stats = db.collection("stats")
+        stats.updateOne({app:"files"}, {$inc:{downloads:1}})
+    }
+    const updateuploads = async ()=>{
+        let stats = db.collection("stats")
+        stats.updateOne({app:"files"}, {$inc:{uploads:1}})
+    }
     const storage = multer.diskStorage({
         destination: function (req, file, cb) {
-            let randstr = randomstring(15)
-            fs.mkdir("./filesdb/" + randstr, () => { })
-            cb(null, './filesdb/' + randstr)
+            cb(null, './filesdb/')
         },
         filename: function (req, file, cb) {
-            cb(null, file.originalname)
+            let randstr = randomstring(15)
+            cb(null, randstr)
         }
     })
 
@@ -39,14 +59,18 @@ module.exports = async function (app) {
 
     function removefiledata(filestr, filesdb) {
         setTimeout(() => {
-            fs.rm("./filesdb/" + filestr, { recursive: true, force: true }, () => { })
-            filesdb.deleteOne({ filestr: filestr })
-        }, 1000 * 60 * 120)
+            filesdb.findOne({ filestr: filestr }).then((data) => {
+                if (data) {
+                    fs.rm("./filesdb/" + filestr, { recursive: true, force: true }, () => { })
+                    filesdb.deleteOne({ filestr: filestr })
+                }
+            })
+        }, 1000 * 60 * process.env.FILES_MINUTES)
     }
 
-    async function addfiledatatodb(fileid, filename, deleteondownload, filestr) {
+    async function addfiledatatodb(fileid, filename, deleteondownload, filestr, size) {
         let files = db.collection("files")
-        files.insertOne({ id: fileid, filename: filename, deleteondownload: deleteondownload, filestr: filestr })
+        files.insertOne({ id: fileid, filename: filename, deleteondownload: deleteondownload, filestr: filestr, size: size })
         removefiledata(filestr, files)
     }
 
@@ -58,22 +82,51 @@ module.exports = async function (app) {
     }
 
 
-    app.post("/files/upload", upload.single("file"), function (req, res) {
+    app.post("/files/cverify", codelimiter, (req, res) => {
+        let ccode = req.body.ccode;
+        if (!ccode) {
+            res.json({ status: false, message: "No Code Provided" })
+            return;
+        }
+        let regexp = /^[0-9]{4}$/
+        if (!regexp.test(ccode)) {
+            res.json({ status: false, message: "Invalid Code" })
+            return;
+        }
+        getfiledata(parseInt(ccode)).then((data) => {
+            if (data) {
+                res.json({ status: false, message: "Code Exists" })
+            } else {
+                res.json({ status: true, message: "Code Available" })
+            }
+        })
+    })
+
+
+    app.post("/files/upload", uploadlimiter, upload.single("file"), function (req, res) {
         try {
-            let filestr = req.file?.destination.split("/")[2]
+            let filestr = req.file?.filename;
             let fileid = getfilesrandomid();
             if (!filestr) {
                 res.json({ status: false, message: "No File Found" })
                 return;
             }
+            let filename = req.file.originalname
             let deleteondownload = req.body.deleteondownload;
-            if (!deleteondownload) {
+            if (!deleteondownload || (deleteondownload !== "false" && deleteondownload !== "true")) {
                 res.json({ status: false, message: "No Deleteondownload Data" })
                 fs.rm("./filesdb/" + filestr, { recursive: true, force: true }, () => { })
                 return;
             }
+            let size = req.body.size;
+            let exp = /^[0-9]{1,2}.[0-9]{2} MB$/
+            if (!size || !exp.test(size)) {
+                res.json({ status: false, message: "No Size Data" })
+                fs.rm("./filesdb/" + filestr, { recursive: true, force: true }, () => { })
+                return;
+            }
             let customstatus = req.body.customstatus;
-            if (!customstatus) {
+            if (!customstatus || (customstatus !== "false" && customstatus !== "true")) {
                 res.json({ status: false, message: "No Custom Id Data" })
                 fs.rm("./filesdb/" + filestr, { recursive: true, force: true }, () => { })
                 return;
@@ -91,50 +144,51 @@ module.exports = async function (app) {
                     fs.rm("./filesdb/" + filestr, { recursive: true, force: true }, () => { })
                     return;
                 }
-                getfiledata(parseInt(customid)).then((data) => {
-                    if (data) {
-                        res.json({ status: false, message: "Custom Id Already Used" })
-                        fs.rm("./filesdb/" + filestr, { recursive: true, force: true }, () => { })
-                        return;
-                    } else {
-                        fileid = parseInt(customid)
-                        let filename = req.file.originalname
-                        addfiledatatodb(fileid, filename, req.body.deleteondownload, filestr)
-                        res.json({ status: true, id: fileid, str: filestr })
-                    }
-                })
-            } else {
-                let filename = req.file.originalname
-                addfiledatatodb(fileid, filename, req.body.deleteondownload, filestr)
-                res.json({ status: true, id: fileid, str: filestr })
+                fileid = parseInt(customid)
             }
-
+            getfiledata(fileid).then((data) => {
+                if (data) {
+                    res.json({ status: false, message: "Custom Id Already Used" })
+                    fs.rm("./filesdb/" + filestr, { recursive: true, force: true }, () => { })
+                } else {
+                    addfiledatatodb(fileid, filename, deleteondownload, filestr, size)
+                    res.json({ status: true, id: fileid, str: filestr })
+                    updateuploads()
+                }
+            })
         } catch (err) {
             fs.rm("./filesdb/" + filestr, { recursive: true, force: true }, () => { })
             res.json({ status: false, message: "Some Error Occuerd" })
         }
     })
 
-    app.get("/files/download/:id", (req, res) => {
+    app.get("/files/download/:id", getfilelimiter, (req, res) => {
+        if (req.headers.range) {
+            return res.status(416).send("Range requests are not supported");
+        }
         let id = req.params.id;
         getfiledata(id).then((data) => {
             if (data) {
-                res.download("./filesdb/" + id + "/" + data.filename, (err) => {
+                let filepath = path.resolve(__dirname, "filesdb", id);
+                res.download(filepath, data.filename, (err) => {
                     if (err) {
-                        res.send({ err: err, msg: "Error Occured" })
+                        console.log("err:" + err);
+                        res.json({ status: false, message: "File Not Found" })
                         return
                     } else if (data.deleteondownload === "true") {
                         fs.rm("./filesdb/" + id, { recursive: true, force: true }, () => { })
+                        let files = db.collection("files")
+                        files.deleteOne({ filestr: data.filestr })
                     }
                 })
             } else {
                 //add the link after creating frontend
-                res.redirect("https://saiteja.fun")
+                res.redirect("http://localhost:5173/?error=File Not Found")
             }
         })
     })
 
-    app.post("/files/download", (req, res) => {
+    app.post("/files/download", downloadlimiter, (req, res) => {
         let token = req.body.token;
         if (!token) {
             res.json({ status: false, message: "Invalid Token" })
@@ -153,7 +207,8 @@ module.exports = async function (app) {
         }
         getfiledata(parseInt(data.id)).then((data) => {
             if (data) {
-                res.json({ status: true, redirect: "/files/download/" + data.filestr })
+                res.json({ status: true, redirect: ("/files/download/" + data.filestr), name: data.filename, size: data.size })
+                updatedownloads()
             } else {
                 res.json({ status: false, message: "No File Found" })
             }
